refactor: ui관련 intersectionObserver훅을 useScrollUITrigger라는 하나의 훅으로 분리
